+++ crates/miden-tx/src/tests/kernel_tests/test_account.rs
@@ -1,7 +1,8 @@
 use miden_lib::{
     errors::tx_kernel_errors::{
         ERR_ACCOUNT_ID_EPOCH_MUST_BE_LESS_THAN_U16_MAX,
-        ERR_ACCOUNT_ID_LEAST_SIGNIFICANT_BYTE_MUST_BE_ZERO, ERR_ACCOUNT_ID_UNKNOWN_STORAGE_MODE,
+        ERR_ACCOUNT_ID_LEAST_SIGNIFICANT_BYTE_MUST_BE_ZERO,
+        ERR_ACCOUNT_ID_NON_PUBLIC_NETWORK_ACCOUNT, ERR_ACCOUNT_ID_UNKNOWN_STORAGE_MODE,
         ERR_ACCOUNT_ID_UNKNOWN_VERSION, TX_KERNEL_ERRORS,
     },
     transaction::TransactionKernel,
@@ -125,6 +126,11 @@ pub fn test_account_validate_id() -> anyhow::Result<()> {
         (ACCOUNT_ID_PUBLIC_FUNGIBLE_FAUCET, None),
         (ACCOUNT_ID_PRIVATE_NON_FUNGIBLE_FAUCET, None),
         (
+            // Set network flag to true while storage mode is public.
+            ACCOUNT_ID_REGULAR_PUBLIC_ACCOUNT_IMMUTABLE_CODE | (1 << 98),
+            None,
+        ),
+        (
             // Set version to a non-zero value (10).
             ACCOUNT_ID_REGULAR_PRIVATE_ACCOUNT_UPDATABLE_CODE | (0x0a << 64),
             Some(ERR_ACCOUNT_ID_UNKNOWN_VERSION),
@@ -138,6 +144,11 @@ pub fn test_account_validate_id() -> anyhow::Result<()> {
             // Set storage mode to an unknown value (0b01).
             ACCOUNT_ID_REGULAR_PRIVATE_ACCOUNT_UPDATABLE_CODE | (0b01 << (64 + 6)),
             Some(ERR_ACCOUNT_ID_UNKNOWN_STORAGE_MODE),
+        ),
+        (
+            // Set network flag to true while storage mode is private.
+            ACCOUNT_ID_REGULAR_PRIVATE_ACCOUNT_UPDATABLE_CODE | (1 << 98),
+            Some(ERR_ACCOUNT_ID_NON_PUBLIC_NETWORK_ACCOUNT),
         ),
         (
             // Set lower 8 bits to a non-zero value (1).
@@ -180,7 +191,7 @@ pub fn test_account_validate_id() -> anyhow::Result<()> {
                 }
             },
             (Err(err), None) => {
-                anyhow::bail!("validation is supposed to succeed but error occurred {}", err)
+                anyhow::bail!("validation is supposed to succeed but error occurred: {}", err)
             },
             (Err(err), Some(_)) => {
                 anyhow::bail!("unexpected different error than expected {}", err)