+++ crates/miden-tx/src/tests/kernel_tests/test_tx.rs
@@ -2,7 +2,9 @@ use alloc::{string::String, sync::Arc, vec::Vec};
 
 use miden_lib::{
     errors::tx_kernel_errors::{
-        ERR_NON_FUNGIBLE_ASSET_ALREADY_EXISTS, ERR_TX_NUMBER_OF_OUTPUT_NOTES_EXCEEDS_LIMIT,
+        ERR_NON_FUNGIBLE_ASSET_ALREADY_EXISTS,
+        ERR_NOTE_NETWORK_EXECUTION_DOES_NOT_TARGET_NETWORK_ACCOUNT,
+        ERR_TX_NUMBER_OF_OUTPUT_NOTES_EXCEEDS_LIMIT,
     },
     transaction::{
         TransactionKernel,
@@ -24,8 +26,9 @@ use miden_objects::{
     },
     testing::{
         account_id::{
-            ACCOUNT_ID_PUBLIC_FUNGIBLE_FAUCET, ACCOUNT_ID_PUBLIC_FUNGIBLE_FAUCET_2,
-            ACCOUNT_ID_REGULAR_PUBLIC_ACCOUNT_IMMUTABLE_CODE,
+            ACCOUNT_ID_NETWORK_NON_FUNGIBLE_FAUCET, ACCOUNT_ID_PRIVATE_FUNGIBLE_FAUCET,
+            ACCOUNT_ID_PRIVATE_SENDER, ACCOUNT_ID_PUBLIC_FUNGIBLE_FAUCET,
+            ACCOUNT_ID_PUBLIC_FUNGIBLE_FAUCET_2, ACCOUNT_ID_REGULAR_PUBLIC_ACCOUNT_IMMUTABLE_CODE,
         },
         constants::NON_FUNGIBLE_ASSET_DATA_2,
     },
@@ -284,7 +287,7 @@ fn test_create_note_too_many_notes() {
             call.wallet::create_note
         end
         ",
-        tag = Felt::new(4),
+        tag = NoteTag::for_local_use_case(1234, 5678).unwrap(),
         recipient = word_to_masm_push_string(&[ZERO, ONE, Felt::new(2), Felt::new(3)]),
         execution_hint_always = Felt::from(NoteExecutionHint::always()),
         PUBLIC_NOTE = NoteType::Public as u8,
@@ -311,9 +314,14 @@ fn test_get_output_notes_commitment() {
     let input_note_2 = tx_context.tx_inputs().input_notes().get_note(1).note();
     let input_asset_2 = **input_note_2.assets().iter().take(1).collect::<Vec<_>>().first().unwrap();
 
+    // Choose random accounts as the target for the note tag.
+    let network_account = AccountId::try_from(ACCOUNT_ID_NETWORK_NON_FUNGIBLE_FAUCET).unwrap();
+    let local_account = AccountId::try_from(ACCOUNT_ID_PRIVATE_FUNGIBLE_FAUCET).unwrap();
+
     // create output note 1
     let output_serial_no_1 = [Felt::new(8); 4];
-    let output_tag_1 = 8888.into();
+    let output_tag_1 =
+        NoteTag::from_account_id(network_account, NoteExecutionMode::Network).unwrap();
     let assets = NoteAssets::new(vec![input_asset_1]).unwrap();
     let metadata = NoteMetadata::new(
         tx_context.tx_inputs().account().id(),
@@ -329,7 +337,7 @@ fn test_get_output_notes_commitment() {
 
     // create output note 2
     let output_serial_no_2 = [Felt::new(11); 4];
-    let output_tag_2 = 1111.into();
+    let output_tag_2 = NoteTag::from_account_id(local_account, NoteExecutionMode::Local).unwrap();
     let assets = NoteAssets::new(vec![input_asset_2]).unwrap();
     let metadata = NoteMetadata::new(
         tx_context.tx_inputs().account().id(),
@@ -466,7 +474,7 @@ fn test_create_note_and_add_asset() {
     let faucet_id = AccountId::try_from(ACCOUNT_ID_PUBLIC_FUNGIBLE_FAUCET).unwrap();
     let recipient = [ZERO, ONE, Felt::new(2), Felt::new(3)];
     let aux = Felt::new(27);
-    let tag = Felt::new(4);
+    let tag = NoteTag::from_account_id(faucet_id, NoteExecutionMode::Local).unwrap();
     let asset = [Felt::new(10), ZERO, faucet_id.suffix(), faucet_id.prefix().as_felt()];
 
     let code = format!(
@@ -536,7 +544,7 @@ fn test_create_note_and_add_multiple_assets() {
 
     let recipient = [ZERO, ONE, Felt::new(2), Felt::new(3)];
     let aux = Felt::new(27);
-    let tag = Felt::new(4);
+    let tag = NoteTag::from_account_id(faucet_2, NoteExecutionMode::Local).unwrap();
 
     let asset = [Felt::new(10), ZERO, faucet.suffix(), faucet.prefix().as_felt()];
     let asset_2 = [Felt::new(20), ZERO, faucet_2.suffix(), faucet_2.prefix().as_felt()];
@@ -637,7 +645,7 @@ fn test_create_note_and_add_same_nft_twice() {
     let tx_context = TransactionContextBuilder::with_standard_account(ONE).build();
 
     let recipient = [ZERO, ONE, Felt::new(2), Felt::new(3)];
-    let tag = Felt::new(4);
+    let tag = NoteTag::for_public_use_case(999, 777, NoteExecutionMode::Local).unwrap();
     let non_fungible_asset = NonFungibleAsset::mock(&[1, 2, 3]);
     let encoded = Word::from(non_fungible_asset);
 
@@ -700,7 +708,7 @@ fn test_build_recipient_hash() {
     // create output note
     let output_serial_no = [ZERO, ONE, Felt::new(2), Felt::new(3)];
     let aux = Felt::new(27);
-    let tag = 8888;
+    let tag = NoteTag::for_public_use_case(42, 42, NoteExecutionMode::Network).unwrap();
     let single_input = 2;
     let inputs = NoteInputs::new(vec![Felt::new(single_input)]).unwrap();
     let input_commitment = inputs.commitment();
@@ -776,6 +784,55 @@ fn test_build_recipient_hash() {
         ),
         recipient_digest.as_slice(),
         "recipient hash not correct",
+    );
+}
+
+#[test]
+fn test_create_network_note_fails_when_target_is_non_public_account() {
+    let tx_context = TransactionContextBuilder::with_standard_account(ONE).build();
+
+    let target_account_id = AccountId::try_from(ACCOUNT_ID_PRIVATE_SENDER).unwrap();
+    let recipient = [ZERO, ONE, Felt::new(2), Felt::new(3)];
+    let aux = Felt::new(42);
+
+    // Manually construct the invalid note tag, because the NoteTag type constructors would return
+    // an error.
+    let prefix_id: u64 = target_account_id.prefix().as_u64();
+    // This results in a note tag equivalent to `NoteTag::from_account_id`. It start with 0b00 and
+    // uses the 30 highest bits from the account ID.
+    let tag = NoteTag::from((prefix_id >> 34) as u32);
+
+    let code = format!(
+        "
+        use.miden::contracts::wallets::basic->wallet
+
+        use.kernel::prologue
+
+        begin
+            exec.prologue::prepare_transaction
+
+            push.{recipient}
+            push.{NOTE_EXECUTION_HINT}
+            push.{PUBLIC_NOTE}
+            push.{aux}
+            push.{tag}
+
+            call.wallet::create_note
+            # => [note_idx]
+
+            # truncate the stack
+            swapdw dropw dropw
+        end
+        ",
+        recipient = word_to_masm_push_string(&recipient),
+        PUBLIC_NOTE = NoteType::Public as u8,
+        NOTE_EXECUTION_HINT = Felt::from(NoteExecutionHint::always()),
+        tag = tag,
+    );
+
+    assert_execution_error!(
+        tx_context.execute_code(&code),
+        ERR_NOTE_NETWORK_EXECUTION_DOES_NOT_TARGET_NETWORK_ACCOUNT
     );
 }